Handoff back to Bryan to make installer work.
bk: 402c1b250qFpY43EM368ghvFDGuI4A

+++ src/utils/extractor.c
@@ -16,10 +16,6 @@
 
 extern unsigned int sfio_size;
 extern unsigned char sfio_data[];
-extern unsigned int shell_size;
-extern unsigned char shell_data[];
-extern unsigned int installer_size;
-extern unsigned char installer_data[];
 extern unsigned int data_size;
 extern unsigned char data_data[];
 
@@ -28,52 +24,57 @@ char * extract(char *, char *, unsigned int, char *);
 main(int ac, char **av)
 {
 	char	*sfio_name;
-	char	*shell_name;
-	char	*installer_name;
 	char	*data_name;
 	char	install_tmp[MAXPATH];
 	char	cmd[4096];
-	int	fd;
+	int	i, fd;
 	pid_t	pid = getpid();
 
 	fprintf(stderr, "Please wait while we unpack the installer...");
+
 	sprintf(install_tmp, "/tmp/bk_install%-u", pid);
 	if (mkdir(install_tmp, 0700) && (errno != EEXIST)) {
 		perror(install_tmp);
 		exit(1);
 	}
 	chdir(install_tmp);
+
 	sfio_name = extract("sfio", sfio_data, sfio_size, install_tmp);
-	shell_name = extract("shell.sfio", shell_data, shell_size, install_tmp);
-	installer_name = extract("installer",
-				installer_data, installer_size, install_tmp);
 	data_name = extract("data", data_data, data_size, install_tmp);
 
+	/*
+	 * Unpack the sfio file, this creates ./bitkeeper/
+	 */
+	sprintf(cmd, "%s -iqm < %s", sfio_name, data_name);
+	system(cmd);
+	if (chdir("bitkeeper")) {
+		perror("bitkeeper");
+		exit(1);
+	}
+	fprintf(stderr, "done.\n");
+	
+	/*
+	 * Run the installer.
+	 */
+	sprintf(cmd, "./bk installtool");
+	for (i = 1; av[i]; i++) {
+		strcat(cmd, " ");
+		strcat(cmd, av[i]);
+	}
+	av[i] = 0;
+	system(cmd);
 
 	/*
-	 * Unpack the shell.sfio file
+	 * Clean up.
 	 */
-	sprintf(cmd, "%s -iqm < %s", sfio_name, shell_name);
-	system(cmd);
-	fprintf(stderr, "done.\n");
-
-	sprintf(cmd, "TCL_LIBRARY=%s/lib/tcl8.4", install_tmp);
-	putenv(strdup(cmd));
-	sprintf(cmd, "%s/gui/bin/tclsh %s", install_tmp, installer_name);
-	for (fd = 1; av[fd]; fd++) {
-		strcat(cmd, " ");
-		strcat(cmd, av[fd]);
+	unless (getenv("BK_SAVE_INSTALL")) {
+		sprintf(cmd, "/bin/rm -rf %s", install_tmp); 
+		system(cmd);	/* careful */
 	}
-	strcat(cmd, " ");
-	strcat(cmd, installer_name);	// XXX - this is not needed
-	strcat(cmd, " ");
-	strcat(cmd, data_name);
-	system(cmd);
-	chdir("..");
-	// XXX - we can't depend on /bin/rm, that's going to have be a bk
-	// command.
-	sprintf(cmd, "/bin/rm -rf %s", install_tmp); 
-	system(cmd);	/* careful */
+
+	/*
+	 * Bitchin'
+	 */
 	exit(0);
 }
 
@@ -92,7 +93,7 @@ extract(char *name, char *x_data, unsigned int x_size, char *install_tmp)
 		exit(1);
 	}
 	if (write(fd, x_data, x_size) != x_size) {
-		perror("write on shell");
+		perror("write");
 		unlink(path);
 		exit(1);
 	}
